Move out `mock_storage.go` to a separate module
